requirements.txt
pandas
qrcode
Pillow
streamlit
openpyxl
